refactor: 重新设计向导组件结构
