+++ src/BootstrapBlazor.Server/Components/Samples/Guide/AnimatedTour.razor
@@ -1,9 +1,5 @@
-﻿@page "/focus-guide"
-@inject IStringLocalizer<FocusGuides> Localizer
-
-<h3>@Localizer["FocusGuideTitle"]</h3>
-
-<h4>@Localizer["FocusGuideSubtitle"]</h4>
+﻿@page "/guide/animated-tour"
+@inject IStringLocalizer<Index> Localizer
 
 <DemoBlock Title="@Localizer["FocusGuideNormalTitle"]"
            Introduction="@Localizer["FocusGuideNormalIntro"]"
