+++ src/BootstrapBlazor.Server/Components/Samples/Guide/Index.razor
@@ -1,0 +1,6 @@
+﻿@page "/guide/index"
+@inject IStringLocalizer<Index> Localizer
+
+<h3>@Localizer["FocusGuideTitle"]</h3>
+
+<h4>@Localizer["FocusGuideSubtitle"]</h4>
