+++ src/BootstrapBlazor.Server/Components/Samples/Guide/Index.razor.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Argo Zhang (argo@163.com). All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+// Website: https://www.blazor.zone or https://argozhang.github.io/
+
+namespace BootstrapBlazor.Server.Components.Samples.Guide;
+
+/// <summary>
+/// FocusGuide 组件示例代码
+/// </summary>
+public partial class Index
+{
+
+}
